requirements.txt
Jinja2>=3.1.0,<4.0.0
Markdown>=3.6,<4.0
PyYAML>=6.0,<7.0
